Added a Home component and the Router
