+++ package.json
@@ -16,6 +16,7 @@
     "@tailwindcss/vite": "^4.1.18",
     "react": "^19.2.0",
     "react-dom": "^19.2.0",
+    "react-router-dom": "^5.3.4",
     "tailwindcss": "^4.1.18"
   },
   "devDependencies": {
